requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
psycopg2